+++ index.html
@@ -109,7 +109,7 @@
             <p>Køn: <span id="dialog-read-member-gender"></span></p>
             <p>Email: <span id="dialog-read-member-email"></span></p>
             <p>Aktivt medlem: <span id="dialog-read-member-membershipActive"></span></p>
-            <p>Medlems siden: <span id="dialog-read-member-membershipDate"></span></p>
+            <p>Medlem siden: <span id="dialog-read-member-membershipDate"></span></p>
             <p>Discipliner: <span id="dialog-read-member-activity"></span></p>
             <p>Medlemstype: <span id="dialog-read-member-membershipType"></span></p>
             <p>Medlemsniveau: <span id="dialog-read-member-membershipLevel"></span></p>
